apps/exports: move mixins to separate files, add docstrings
