+++ adhocracy4/exports/mixins/comments.py
@@ -1,0 +1,66 @@
+from django.utils.translation import ugettext as _
+
+from adhocracy4.comments.models import Comment
+from adhocracy4.exports import unescape_and_strip_html
+
+from .base import VirtualFieldMixin
+
+
+class ItemExportWithCommentCountMixin(VirtualFieldMixin):
+    """
+    Adds the (parent) comment count to an item.
+    """
+    def get_virtual_fields(self, virtual):
+        if 'comment_count' not in virtual:
+            virtual['comment_count'] = _('Comment count')
+        return super().get_virtual_fields(virtual)
+
+    def get_comment_count_data(self, item):
+        # FIXME: the annotated comment_count does currently not include replies
+        # if hasattr(item, 'comment_count'):
+        #     return item.comment_count
+        if hasattr(item, 'comments'):
+            return self._count_comments(item)
+
+        return 0
+
+    def _count_comments(self, item):
+        comment_ids = item.comments.values_list('id', flat=True)
+        replies = Comment.objects.filter(parent_comment__in=comment_ids)
+        return len(comment_ids) + len(replies)
+
+
+class ItemExportWithCommentsMixin(VirtualFieldMixin):
+    """
+    Adds the comments to the item.
+
+    This is meant to be used in an item (not comment) export. It adds
+    all comments and the replies to the same line as the item.
+    """
+    COMMENT_FMT = '{date} - {username}\n{text}'
+    REPLY_FMT = '@reply: {date} - {username}\n{text}'
+
+    def get_virtual_fields(self, virtual):
+        if 'comments' not in virtual:
+            virtual['comments'] = _('Comments')
+        return super().get_virtual_fields(virtual)
+
+    def get_comments_data(self, item):
+        if hasattr(item, 'comments'):
+            return '\n----\n'.join(self._flat_comments(item))
+        return ''
+
+    def _flat_comments(self, item):
+        for comment in item.comments.all():
+            yield self.COMMENT_FMT.format(
+                date=comment.created.isoformat(),
+                username=comment.creator.username,
+                text=unescape_and_strip_html(comment.comment)
+            )
+
+            for reply in comment.child_comments.all():
+                yield self.REPLY_FMT.format(
+                    date=reply.created.isoformat(),
+                    username=reply.creator.username,
+                    text=unescape_and_strip_html(reply.comment)
+                )
